Fix unit test for transactions and transaction pool
