+++ framework/src/modules/chain/transactions/exceptions_handlers.js
